requirements.txt
evdev>=1.6
pyyaml>=6.0
